Move LogInitialization to a common web server module
