+++ modules/edu.gemini.p1backend/edu.gemini.p1backend.server/src/main/scala/edu/gemini/p1backend/server/WebServerLauncher.scala
@@ -3,11 +3,8 @@ package edu.gemini.seqexec.web.server.http4s
 import java.io.File
 import java.util.logging.Logger
 
-import edu.gemini.seqexec.engine
-import edu.gemini.seqexec.model.Model.SeqexecEvent
-import edu.gemini.seqexec.server.SeqexecEngine
-import edu.gemini.seqexec.web.server.common.LogInitialization
-import edu.gemini.seqexec.web.server.security.{AuthenticationConfig, AuthenticationService, LDAPConfig}
+import edu.gemini.web.server.common.LogInitialization
+
 import knobs._
 import org.http4s.server.{Server, ServerApp}
 import org.http4s.server.blaze.BlazeBuilder
@@ -51,46 +48,16 @@ object WebServerLauncher extends ServerApp with LogInitialization {
       WebServerConfiguration(host, port, devMode.equalsIgnoreCase("dev"))
     }
 
-  // Configuration of the ldap clients
-  val ldapConf: Task[LDAPConfig] =
-    config.map { cfg =>
-      val urls = cfg.require[List[String]]("authentication.ldapURLs")
-      LDAPConfig(urls)
-    }
-
-  // Configuration of the authentication service
-  val authConf: Task[AuthenticationConfig] =
-    for {
-      ld <- ldapConf
-      cfg <- config
-    } yield {
-      val devMode = cfg.require[String]("mode")
-      val sessionTimeout = cfg.require[Int]("authentication.sessionLifeHrs")
-      val cookieName = cfg.require[String]("authentication.cookieName")
-      val secretKey = cfg.require[String]("authentication.secretKey")
-      val useSSL = cfg.require[Boolean]("authentication.useSSL")
-      AuthenticationConfig(devMode.equalsIgnoreCase("dev"), sessionTimeout, cookieName, secretKey, useSSL, ld)
-    }
-
-  /**
-    * Configures the Authentication service
-    */
-  def authService: Kleisli[Task, AuthenticationConfig, AuthenticationService] = Kleisli { conf =>
-    Task.delay(AuthenticationService(conf))
-  }
-
   /**
     * Configures and builds the web server
     */
-  def webServer(as: AuthenticationService, events: (engine.EventQueue, Topic[SeqexecEvent]), se: SeqexecEngine): Kleisli[Task, WebServerConfiguration, Server] = Kleisli { conf =>
+  def webServer: Kleisli[Task, WebServerConfiguration, Server] = Kleisli { conf =>
     val logger = Logger.getLogger(getClass.getName)
     logger.info(s"Start server on ${conf.devMode ? "dev" | "production"} mode")
 
     BlazeBuilder.bindHttp(conf.port, conf.host)
       .withWebSockets(true)
       .mountService(new StaticRoutes(conf.devMode).service, "/")
-      .mountService(new SeqexecCommandRoutes(as, events._1, se).service, "/api/seqexec/commands")
-      .mountService(new SeqexecUIApiRoutes(as, events, se).service, "/api")
       .start
   }
 
@@ -98,28 +65,11 @@ object WebServerLauncher extends ServerApp with LogInitialization {
     * Reads the configuration and launches the web server
     */
   override def server(args: List[String]): Task[Server] = {
-    val engineTask = for {
-      c    <- config
-      seqc <- SeqexecEngine.seqexecConfiguration.run(c)
-    } yield SeqexecEngine(seqc)
-
-    val inq  = async.boundedQueue[engine.Event](10)
-    val out  = async.topic[SeqexecEvent]()
-
-    engineTask flatMap (
-      se => Nondeterminism[Task].both(
-        // Launch engine and broadcast channel
-        se.eventProcess(inq).to(out.publish).run,
-        // Launch web server
-        for {
-          _  <- configLog
-          ac <- authConf
-          as <- authService.run(ac)
-          wc <- serverConf
-          ws <- webServer(as, (inq, out), se).run(wc)
-        } yield ws
-      ).map(_._2)
-    )
+    // Launch web server
+    for {
+      _  <- configLog
+      ws <- webServer
+    } yield ws
   }
 
 }
